requirements.txt
mediapipe==0.10.1
Flask==3.0.0
flask-basicauth==0.2.0
numpy==1.23.5
opencv-python==4.6.0.66
Pillow==9.5.0
scikit-learn==1.3.0
pip~=21.3.1
attrs~=23.1.0
wheel~=0.37.1
tornado~=6.3.3
Jinja2~=3.1.2
setuptools~=60.2.0
packaging~=23.1
pyparsing~=3.0.9
zipp~=3.17.0
joblib~=1.3.1
gunicorn==20.1.0